[#206] rename: PostComment, LikeComment 패키지 이동
