+++ src/main/java/com/konggogi/veganlife/comment/service/CommentLikeQueryService.java
@@ -1,8 +1,8 @@
 package com.konggogi.veganlife.comment.service;
 
 
+import com.konggogi.veganlife.comment.domain.CommentLike;
 import com.konggogi.veganlife.comment.repository.CommentLikeRepository;
-import com.konggogi.veganlife.like.domain.CommentLike;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
