+++ src/main/java/com/konggogi/veganlife/post/service/PostLikeService.java
@@ -2,11 +2,11 @@ package com.konggogi.veganlife.post.service;
 
 
 import com.konggogi.veganlife.global.exception.ErrorCode;
-import com.konggogi.veganlife.like.domain.PostLike;
 import com.konggogi.veganlife.like.domain.mapper.LikeMapper;
 import com.konggogi.veganlife.member.domain.Member;
 import com.konggogi.veganlife.member.service.MemberQueryService;
 import com.konggogi.veganlife.post.domain.Post;
+import com.konggogi.veganlife.post.domain.PostLike;
 import com.konggogi.veganlife.post.exception.IllegalLikeStatusException;
 import com.konggogi.veganlife.post.repository.PostLikeRepository;
 import lombok.RequiredArgsConstructor;
